Extract CensusPanel, NeighborsPanel, ScatterSection from +layout.svelte
Break up the 2450-line agency layout into focused components. Each
component manages its own expanded/toggle state and imports only the
translations it needs.

closes #151
