+++ packages/web/src/routes/agency/[slug]/+layout.svelte
@@ -1,11 +1,13 @@
 <script>
   import { Grid, PagingData, PlainTableCssTheme } from "@mediakular/gridcraft";
   import AgencyMap from "$lib/components/AgencyMap.svelte";
-  import AgencyRateScatter from "$lib/components/AgencyRateScatter.svelte";
   import GridMetricCell from "$lib/components/grid/GridMetricCell.svelte";
   import GridTextCell from "$lib/components/grid/GridTextCell.svelte";
   import GridValueCell from "$lib/components/grid/GridValueCell.svelte";
   import MetricChartModal from "$lib/components/MetricChartModal.svelte";
+  import CensusPanel from "$lib/components/CensusPanel.svelte";
+  import NeighborsPanel from "$lib/components/NeighborsPanel.svelte";
+  import ScatterSection from "$lib/components/ScatterSection.svelte";
   import StickyHeader from "$lib/components/StickyHeader.svelte";
   import { goto } from "$app/navigation";
   import { page } from "$app/stores";
@@ -31,31 +33,9 @@
     race_other,
     race_total,
     race_white,
-    census_heading,
-    census_toggle_hide,
-    census_toggle_show,
-    census_source_attribution,
     census_stat_population,
     census_stat_median_age,
     census_stat_median_income,
-    census_race_heading,
-    census_no_race_data,
-    census_location_heading,
-    census_congressional_district,
-    census_congressional_districts,
-    census_state_legislative_district,
-    census_state_legislative_districts,
-    census_metro_area_label,
-    census_acs_tables_heading,
-    census_acs_table_row_count,
-    census_acs_col_category,
-    census_acs_col_value,
-    census_acs_col_pct,
-    census_acs_col_moe,
-    census_no_acs_data,
-    census_acs_survey_note_full,
-    census_acs_survey_note_brief,
-    census_geography_note,
   } from "$lib/paraglide/messages";
 
   /** @type {import('./$types').LayoutData} */
@@ -111,8 +91,6 @@
   let basemapStyleUrl = "/map/style.en.json";
   let metricSearchTimeout;
   let lastTrackedMetricSearch = "";
-  let neighborsExpanded = false;
-  let censusExpanded = false;
   let agencyComments = [];
   let selectedAgencyComment = null;
   let selectedCommentText = "";
@@ -125,25 +103,6 @@
   const trackEvent = (event, payload = {}) => {
     if (typeof window === "undefined") return;
     window.umami?.track?.(event, payload);
-  };
-
-  const toggleNeighbors = () => {
-    const next = !neighborsExpanded;
-    neighborsExpanded = next;
-    trackEvent(next ? "agency_neighbors_expand" : "agency_neighbors_collapse", {
-      agency: agencyData?.agency ?? data.slug,
-      touchingCount: touchingAgencies.length,
-      containedCount: containedAgencies.length,
-    });
-  };
-
-  const toggleCensus = () => {
-    const next = !censusExpanded;
-    censusExpanded = next;
-    trackEvent(next ? "agency_census_expand" : "agency_census_collapse", {
-      agency: agencyData?.agency ?? data.slug,
-      hasJurisdictionData: Boolean(geocodioDemographics),
-    });
   };
 
   $: agencyData = data.data;
@@ -1513,298 +1472,17 @@
     />
   </section>
 
-  {#if touchingAgencies.length || containedAgencies.length}
-    <section class="mb-6">
-      <div class="rounded-2xl border border-slate-200 bg-white p-4">
-        <button
-          type="button"
-          class="flex w-full items-center gap-4 text-left"
-          aria-expanded={neighborsExpanded}
-          aria-controls="neighbors-panel"
-          on:click={toggleNeighbors}
-        >
-          <span
-            class={`inline-flex h-9 w-9 shrink-0 items-center justify-center rounded-full border border-slate-200 text-slate-500 transition ${
-              neighborsExpanded ? "rotate-180 bg-slate-100" : "bg-white"
-            }`}
-          >
-            <svg
-              aria-hidden="true"
-              class="h-4 w-4"
-              viewBox="0 0 20 20"
-              fill="none"
-              stroke="currentColor"
-              stroke-width="2"
-              stroke-linecap="round"
-              stroke-linejoin="round"
-            >
-              <path d="M6 8l4 4 4-4" />
-            </svg>
-            <span class="sr-only">
-              {neighborsExpanded ? "Hide" : "Show"}
-            </span>
-          </span>
-          <div>
-            <h2 class="text-lg font-semibold text-slate-900 sm:text-xl">
-              {m?.agency_neighbors_heading?.() ?? "Nearby agencies"}
-            </h2>
-            <p class="mt-1 text-xs text-slate-500">
-              {touchingAgencies.length} touching • {containedAgencies.length} contained
-            </p>
-          </div>
-        </button>
-        {#if neighborsExpanded}
-          <div
-            id="neighbors-panel"
-            class="mt-4 space-y-4 border-t border-slate-100 pt-4"
-          >
-            {#if touchingAgencies.length}
-              <div>
-                <h3 class="text-xs font-semibold uppercase tracking-[0.2em] text-slate-400">
-                  {m?.agency_neighbors_touching_label?.() ?? "Touching agencies"}
-                </h3>
-                <div class="mt-2 flex flex-wrap gap-x-3 gap-y-1 text-sm text-slate-700">
-                  {#each touchingAgencies as neighbor}
-                    <a class="underline" href={`${localeBase}/agency/${neighbor.slug}`}>
-                      {neighbor.label}
-                    </a>
-                  {/each}
-                </div>
-              </div>
-            {/if}
-            {#if containedAgencies.length}
-              <div>
-                <h3 class="text-xs font-semibold uppercase tracking-[0.2em] text-slate-400">
-                  {m?.agency_neighbors_contained_label?.() ?? "Contained agencies"}
-                </h3>
-                <div class="mt-2 flex flex-wrap gap-x-3 gap-y-1 text-sm text-slate-700">
-                  {#each containedAgencies as neighbor}
-                    <a class="underline" href={`${localeBase}/agency/${neighbor.slug}`}>
-                      {neighbor.label}
-                    </a>
-                  {/each}
-                </div>
-              </div>
-            {/if}
-          </div>
-        {/if}
-      </div>
-    </section>
-  {/if}
+  <NeighborsPanel
+    {touchingAgencies}
+    {containedAgencies}
+    {localeBase}
+    agencySlug={agencyData?.agency ?? data.slug}
+  />
 
-  {#if geocodioDemographics}
-    <section class="mb-10">
-      <div class="rounded-2xl border border-slate-200 bg-white p-4">
-        <button
-          type="button"
-          class="flex w-full items-center gap-4 text-left"
-          aria-expanded={censusExpanded}
-          aria-controls="census-panel"
-          on:click={toggleCensus}
-        >
-          <span
-            class={`inline-flex h-9 w-9 shrink-0 items-center justify-center rounded-full border border-slate-200 text-slate-500 transition ${
-              censusExpanded ? "rotate-180 bg-slate-100" : "bg-white"
-            }`}
-          >
-            <svg
-              aria-hidden="true"
-              class="h-4 w-4"
-              viewBox="0 0 20 20"
-              fill="none"
-              stroke="currentColor"
-              stroke-width="2"
-              stroke-linecap="round"
-              stroke-linejoin="round"
-            >
-              <path d="M6 8l4 4 4-4" />
-            </svg>
-            <span class="sr-only">
-              {censusExpanded ? census_toggle_hide() : census_toggle_show()}
-            </span>
-          </span>
-          <div>
-            <h2 class="text-lg font-semibold text-slate-900 sm:text-xl">
-              {census_heading()}
-            </h2>
-          </div>
-        </button>
-        {#if censusExpanded}
-          <div id="census-panel" class="mt-4 space-y-4 border-t border-slate-100 pt-4">
-            <article class="rounded-xl border border-slate-300 bg-slate-50/50 p-4 sm:p-5">
-              <div class="flex flex-wrap items-start justify-between gap-2">
-                <p class="text-sm text-slate-700">
-                  {census_source_attribution()}
-                </p>
-                {#if geocodioDemographics.formattedAddress}
-                  <p class="text-sm text-slate-600">{geocodioDemographics.formattedAddress}</p>
-                {/if}
-              </div>
-
-              {#if geocodioDemographics.summaryStats.length}
-                <div class="mt-4 grid gap-2 sm:grid-cols-3">
-                  {#each geocodioDemographics.summaryStats as stat}
-                    <div class="rounded-lg border border-slate-300 bg-white p-3">
-                      <p class="text-xs font-semibold uppercase tracking-[0.12em] text-slate-600">
-                        {stat.label}
-                      </p>
-                      <p class="mt-1 text-2xl font-semibold text-slate-900 sm:text-3xl">
-                        {stat.value}
-                      </p>
-                    </div>
-                  {/each}
-                </div>
-              {/if}
-
-              <div class="mt-5 grid gap-3 md:grid-cols-2 md:items-start">
-                <div class="rounded-lg border border-slate-300 bg-white p-3">
-                  <h4 class="text-sm font-semibold uppercase tracking-[0.12em] text-slate-700">
-                    {census_race_heading()}
-                  </h4>
-                  {#if geocodioDemographics.raceRows.length}
-                    <div class="mt-2 grid gap-y-2 text-sm text-slate-800">
-                      {#each geocodioDemographics.raceRows as row}
-                        <div class="flex items-baseline justify-between gap-3">
-                          <span>{row.label}</span>
-                          <span class="font-mono tabular-nums text-slate-900">{row.display}</span>
-                        </div>
-                      {/each}
-                    </div>
-                  {:else}
-                    <p class="mt-2 text-sm text-slate-700">
-                      {census_no_race_data()}
-                    </p>
-                  {/if}
-                </div>
-
-                <div class="rounded-lg border border-slate-300 bg-white p-3">
-                  <h4 class="text-sm font-semibold uppercase tracking-[0.12em] text-slate-700">
-                    {census_location_heading()}
-                  </h4>
-                  <dl class="mt-2 space-y-3 text-sm text-slate-800">
-                    <div>
-                      <dt class="font-medium text-slate-600">
-                        {geocodioDemographics.congressionalDistricts.length > 1
-                          ? census_congressional_districts()
-                          : census_congressional_district()}
-                      </dt>
-                      <dd class="mt-0.5">
-                        {#if geocodioDemographics.congressionalDistricts.length}
-                          {#each geocodioDemographics.congressionalDistricts as district, index}
-                            <span>{district}{index < geocodioDemographics.congressionalDistricts.length - 1 ? ", " : ""}</span>
-                          {/each}
-                        {:else}
-                          —
-                        {/if}
-                      </dd>
-                    </div>
-                    <div>
-                      <dt class="font-medium text-slate-600">
-                        {geocodioDemographics.stateLegislativeDistricts.length > 1
-                          ? census_state_legislative_districts()
-                          : census_state_legislative_district()}
-                      </dt>
-                      <dd class="mt-0.5">
-                        {#if geocodioDemographics.stateLegislativeDistricts.length}
-                          {#each geocodioDemographics.stateLegislativeDistricts as district, index}
-                            <span>{district}{index < geocodioDemographics.stateLegislativeDistricts.length - 1 ? ", " : ""}</span>
-                          {/each}
-                        {:else}
-                          —
-                        {/if}
-                      </dd>
-                    </div>
-                    <div>
-                      <dt class="font-medium text-slate-600">{census_metro_area_label()}</dt>
-                      <dd class="mt-0.5">{geocodioDemographics.metroAreaName || "—"}</dd>
-                    </div>
-                  </dl>
-                </div>
-              </div>
-
-              {#if geocodioDemographics.acsTableGroups.length}
-                <div class="mt-5">
-                  <h4 class="text-sm font-semibold uppercase tracking-[0.12em] text-slate-700">
-                    {census_acs_tables_heading()}
-                  </h4>
-                  <div class="mt-2 space-y-3">
-                    {#each geocodioDemographics.acsTableGroups as group}
-                      <div class="rounded-lg border border-slate-300 bg-white p-3">
-                        <h5 class="text-sm font-semibold text-slate-900">{group.label}</h5>
-                        <div class="mt-2 space-y-2">
-                          {#each group.tables as table}
-                            <details class="rounded-md border border-slate-200 bg-slate-50/70">
-                              <summary class="cursor-pointer list-none px-3 py-2.5">
-                                <div class="flex items-center justify-between gap-3">
-                                  <span class="text-sm font-medium text-slate-900">{table.title}</span>
-                                  <span class="shrink-0 text-xs text-slate-600">{census_acs_table_row_count({ count: table.rowCount })}</span>
-                                </div>
-                              </summary>
-                              <div class="border-t border-slate-200 px-3 pb-3 pt-2">
-                                <div class="mt-1 overflow-x-auto">
-                                  <table class="min-w-full table-auto border-collapse">
-                                    <colgroup>
-                                      <col />
-                                      <col style="width: 7rem;" />
-                                      <col style="width: 5rem;" />
-                                      <col style="width: 6rem;" />
-                                    </colgroup>
-                                    <thead>
-                                      <tr class="text-[0.68rem] font-semibold uppercase tracking-[0.08em] text-slate-600">
-                                        <th class="py-1 text-left">{census_acs_col_category()}</th>
-                                        <th class="py-1 text-right">{census_acs_col_value()}</th>
-                                        <th class="py-1 text-right">{census_acs_col_pct()}</th>
-                                        <th class="py-1 text-right">{census_acs_col_moe()}</th>
-                                      </tr>
-                                    </thead>
-                                    <tbody class="divide-y divide-slate-200">
-                                      {#each table.rows as row}
-                                        <tr class="text-sm text-slate-800">
-                                          <td class="py-1.5 pr-2">{row.label}</td>
-                                          <td class="py-1.5 text-right font-mono tabular-nums text-slate-900">
-                                            {row.valueDisplay || "—"}
-                                          </td>
-                                          <td class="py-1.5 text-right font-mono tabular-nums text-slate-900">
-                                            {row.percentDisplay || "—"}
-                                          </td>
-                                          <td class="py-1.5 text-right font-mono tabular-nums text-slate-500">
-                                            {row.moeDisplay ? `±${row.moeDisplay}` : "—"}
-                                          </td>
-                                        </tr>
-                                      {/each}
-                                    </tbody>
-                                  </table>
-                                </div>
-                              </div>
-                            </details>
-                          {/each}
-                        </div>
-                      </div>
-                    {/each}
-                  </div>
-                </div>
-              {:else if !geocodioDemographics.hasAcs}
-                <p class="mt-4 text-sm text-slate-700">
-                  {census_no_acs_data()}
-                </p>
-              {/if}
-
-              <p class="mt-4 text-xs leading-relaxed text-slate-600">
-                {#if geocodioDemographics.surveyYears}
-                  {geocodioDemographics.surveyDurationYears
-                    ? census_acs_survey_note_full({ years: geocodioDemographics.surveyYears, duration: geocodioDemographics.surveyDurationYears })
-                    : census_acs_survey_note_brief({ years: geocodioDemographics.surveyYears })}
-                {/if}
-                {#if geocodioDemographics.censusYear}
-                  {census_geography_note({ year: geocodioDemographics.censusYear })}
-                {/if}
-              </p>
-            </article>
-          </div>
-        {/if}
-      </div>
-    </section>
-  {/if}
+  <CensusPanel
+    {geocodioDemographics}
+    agencySlug={agencyData?.agency ?? data.slug}
+  />
 
 
   <section class="mb-10">
@@ -1915,417 +1593,14 @@
                 {` ${stopCountFormatter.format(reportingAgencyCount)} agencies reporting in ${selectedYear}.`}
               {/if}
             </p>
-            <div class="mt-12">
-              <div class="flex flex-col gap-3">
-                <div
-                  role="tablist"
-                  aria-label={agency_yearly_data_heading()}
-                  class="mt-12 flex flex-wrap items-center gap-2"
-                >
-                  {#each years as year}
-                    <button
-                      type="button"
-                      role="tab"
-                      aria-selected={year === selectedYear}
-                      class={`rounded-md border px-3 py-1.5 text-sm font-semibold tracking-wide transition sm:text-base ${
-                        year === selectedYear
-                          ? "border-slate-900 bg-slate-900 text-white"
-                          : "border-slate-200 bg-white text-slate-600 hover:border-slate-300 hover:text-slate-900"
-                      }`}
-                      on:click={() => selectYear(year, "bottom")}
-                    >
-                      {year}
-                    </button>
-                  {/each}
-                </div>
-                <div class="relative left-1/2 mt-4 w-screen -translate-x-1/2 space-y-4 px-4 sm:px-6">
-                  <div class="grid gap-7 lg:gap-8 lg:grid-cols-3">
-                    <AgencyRateScatter
-                      selectedYear={selectedYear}
-                      agencyName={agencyData?.agency ?? data.slug}
-                      title="Total stops vs citations"
-                      domainGroup="stops-citations"
-                      showMeanLines={true}
-                      xLabel={m?.agency_scatter_total_stops_label?.() ?? "Total stops"}
-                      yLabel={m?.agency_scatter_citations_label?.() ?? "Citations"}
-                      xMetricKey="rates-by-race--totals--all-stops"
-                      yMetricKey="rates-by-race--totals--citations"
-                      excludeAgencies={scatterExcludedAgencies}
-                      minStops={500}
-                      sizeByStops={true}
-                      stopsLabel={m?.agency_scatter_total_stops_label?.() ?? "Total stops"}
-                      minCount={50}
-                      minCountKey="rates-by-race--totals--citations"
-                      minCountMessage={
-                        m?.agency_scatter_min_citations_note?.() ??
-                          "Requires at least 50 citations to display."
-                      }
-                      note={
-                        m?.agency_scatter_min_citations_note?.() ??
-                          "Requires at least 50 citations to display."
-                      }
-                      xScaleType="log"
-                      yScaleType="log"
-                      minX={1}
-                      minY={1}
-                    />
-                    <AgencyRateScatter
-                      selectedYear={selectedYear}
-                      agencyName={agencyData?.agency ?? data.slug}
-                      title="Non-white stops vs citations"
-                      domainGroup="stops-citations"
-                      showMeanLines={true}
-                      xLabel="Non-white total stops"
-                      yLabel="Non-white citations"
-                      xMetricKey="rates-by-race--totals--all-stops"
-                      yMetricKey="rates-by-race--totals--citations"
-                      xColumn="Non-white"
-                      yColumn="Non-white"
-                      excludeAgencies={scatterExcludedAgencies}
-                      minStops={500}
-                      sizeByStops={true}
-                      stopsColumn="Non-white"
-                      stopsLabel="Non-white total stops"
-                      minCount={25}
-                      minCountKey="rates-by-race--totals--citations"
-                      minCountColumn="Non-white"
-                      minCountMessage={
-                        m?.agency_scatter_min_citations_note_small?.() ??
-                          "Requires at least 25 citations to display."
-                      }
-                      note={
-                        m?.agency_scatter_min_citations_note_small?.() ??
-                          "Requires at least 25 citations to display."
-                      }
-                      xScaleType="log"
-                      yScaleType="log"
-                      minX={1}
-                      minY={1}
-                    />
-                    <AgencyRateScatter
-                      selectedYear={selectedYear}
-                      agencyName={agencyData?.agency ?? data.slug}
-                      title="White stops vs citations"
-                      domainGroup="stops-citations"
-                      showMeanLines={true}
-                      xLabel="White total stops"
-                      yLabel="White citations"
-                      xMetricKey="rates-by-race--totals--all-stops"
-                      yMetricKey="rates-by-race--totals--citations"
-                      xColumn="White"
-                      yColumn="White"
-                      excludeAgencies={scatterExcludedAgencies}
-                      minStops={500}
-                      sizeByStops={true}
-                      stopsColumn="White"
-                      stopsLabel="White total stops"
-                      minCount={25}
-                      minCountKey="rates-by-race--totals--citations"
-                      minCountColumn="White"
-                      minCountMessage={
-                        m?.agency_scatter_min_citations_note_small?.() ??
-                          "Requires at least 25 citations to display."
-                      }
-                      note={
-                        m?.agency_scatter_min_citations_note_small?.() ??
-                          "Requires at least 25 citations to display."
-                      }
-                      xScaleType="log"
-                      yScaleType="log"
-                      minX={1}
-                      minY={1}
-                    />
-                  </div>
-                  <div class="grid gap-7 lg:gap-8 lg:grid-cols-3">
-                    <AgencyRateScatter
-                      selectedYear={selectedYear}
-                      agencyName={agencyData?.agency ?? data.slug}
-                      title={
-                        m?.agency_scatter_citation_vs_arrest_heading?.() ??
-                          "Citation rate vs arrest rate"
-                      }
-                      domainGroup="citation-arrest"
-                      showMeanLines={true}
-                      xLabel={m?.agency_scatter_citation_rate_label?.() ?? "Citation rate (%)"}
-                      yLabel={m?.agency_scatter_arrest_rate_label?.() ?? "Arrest rate (%)"}
-                      xMetricKey="rates-by-race--totals--citations-rate"
-                      yMetricKey="rates-by-race--totals--arrests-rate"
-                      excludeAgencies={scatterExcludedAgencies}
-                      formatXAxisTick={formatPercentTick}
-                      formatYAxisTick={formatPercentTick}
-                      xCountKey="rates-by-race--totals--citations"
-                      yCountKey="rates-by-race--totals--arrests"
-                      xCountLabel={m?.agency_scatter_citations_label?.() ?? "Citations"}
-                      yCountLabel={m?.agency_scatter_arrests_label?.() ?? "Arrests"}
-                      minStops={500}
-                      sizeByStops={true}
-                      stopsLabel={m?.agency_scatter_total_stops_label?.() ?? "Total stops"}
-                      minCount={50}
-                      minCountKey="rates-by-race--totals--citations"
-                      minCountMessage={
-                        m?.agency_scatter_min_citations_note?.() ??
-                          "Requires at least 50 citations to display."
-                      }
-                      note={
-                        m?.agency_scatter_min_citations_note?.() ??
-                          "Requires at least 50 citations to display."
-                      }
-                      excludeExactValue={100}
-                      minX={0}
-                      minY={0}
-                      maxX={100}
-                      maxY={100}
-                    />
-                    <AgencyRateScatter
-                      selectedYear={selectedYear}
-                      agencyName={agencyData?.agency ?? data.slug}
-                      title={
-                        m?.agency_scatter_citation_vs_arrest_nonwhite_heading?.() ??
-                          "Non-white citation rate vs arrest rate"
-                      }
-                      domainGroup="citation-arrest"
-                      showMeanLines={true}
-                      xLabel={
-                        m?.agency_scatter_citation_rate_nonwhite_label?.() ??
-                          "Non-white citation rate"
-                      }
-                      yLabel={
-                        m?.agency_scatter_arrest_rate_nonwhite_label?.() ??
-                          "Non-white arrest rate"
-                      }
-                      xMetricKey="rates-by-race--totals--citations-rate"
-                      yMetricKey="rates-by-race--totals--arrests-rate"
-                      xColumn="Non-white"
-                      yColumn="Non-white"
-                      excludeAgencies={scatterExcludedAgencies}
-                      formatXAxisTick={formatPercentTick}
-                      formatYAxisTick={formatPercentTick}
-                      xCountKey="rates-by-race--totals--citations"
-                      yCountKey="rates-by-race--totals--arrests"
-                      xCountColumn="Non-white"
-                      yCountColumn="Non-white"
-                      xCountLabel={
-                        m?.agency_scatter_citations_nonwhite_label?.() ??
-                          "Non-white citations"
-                      }
-                      yCountLabel={
-                        m?.agency_scatter_arrests_nonwhite_label?.() ??
-                          "Non-white arrests"
-                      }
-                      minStops={500}
-                      sizeByStops={true}
-                      stopsColumn="Non-white"
-                      stopsLabel="Non-white total stops"
-                      minCount={25}
-                      minCountKey="rates-by-race--totals--citations"
-                      minCountColumn="Non-white"
-                      minCountMessage={
-                        m?.agency_scatter_min_citations_note_small?.() ??
-                          "Requires at least 25 citations to display."
-                      }
-                      note={
-                        m?.agency_scatter_min_citations_note_small?.() ??
-                          "Requires at least 25 citations to display."
-                      }
-                      excludeExactValue={100}
-                      minX={0}
-                      minY={0}
-                      maxX={100}
-                      maxY={100}
-                    />
-                    <AgencyRateScatter
-                      selectedYear={selectedYear}
-                      agencyName={agencyData?.agency ?? data.slug}
-                      title={
-                        m?.agency_scatter_citation_vs_arrest_white_heading?.() ??
-                          "White citation rate vs arrest rate"
-                      }
-                      domainGroup="citation-arrest"
-                      showMeanLines={true}
-                      xLabel={
-                        m?.agency_scatter_citation_rate_white_label?.() ??
-                          "White citation rate"
-                      }
-                      yLabel={
-                        m?.agency_scatter_arrest_rate_white_label?.() ??
-                          "White arrest rate"
-                      }
-                      xMetricKey="rates-by-race--totals--citations-rate"
-                      yMetricKey="rates-by-race--totals--arrests-rate"
-                      xColumn="White"
-                      yColumn="White"
-                      excludeAgencies={scatterExcludedAgencies}
-                      formatXAxisTick={formatPercentTick}
-                      formatYAxisTick={formatPercentTick}
-                      xCountKey="rates-by-race--totals--citations"
-                      yCountKey="rates-by-race--totals--arrests"
-                      xCountColumn="White"
-                      yCountColumn="White"
-                      xCountLabel={m?.agency_scatter_citations_white_label?.() ?? "White citations"}
-                      yCountLabel={m?.agency_scatter_arrests_white_label?.() ?? "White arrests"}
-                      minStops={500}
-                      sizeByStops={true}
-                      stopsColumn="White"
-                      stopsLabel="White total stops"
-                      minCount={25}
-                      minCountKey="rates-by-race--totals--citations"
-                      minCountColumn="White"
-                      minCountMessage={
-                        m?.agency_scatter_min_citations_note_small?.() ??
-                          "Requires at least 25 citations to display."
-                      }
-                      note={
-                        m?.agency_scatter_min_citations_note_small?.() ??
-                          "Requires at least 25 citations to display."
-                      }
-                      excludeExactValue={100}
-                      minX={0}
-                      minY={0}
-                      maxX={100}
-                      maxY={100}
-                    />
-                  </div>
-                  <div class="grid gap-7 lg:gap-8 lg:grid-cols-3">
-                    <AgencyRateScatter
-                      selectedYear={selectedYear}
-                      agencyName={agencyData?.agency ?? data.slug}
-                      title={
-                        m?.agency_scatter_search_vs_hit_heading?.() ??
-                          "Search rate vs contraband hit rate"
-                      }
-                      domainGroup="hits-searches"
-                      showMeanLines={true}
-                      xLabel={m?.agency_scatter_search_rate_label?.() ?? "Search rate"}
-                      yLabel={m?.agency_scatter_hit_rate_label?.() ?? "Hit rate"}
-                      xMetricKey="rates-by-race--totals--searches-rate"
-                      yMetricKey="rates-by-race--totals--contraband-hit-rate"
-                      formatXAxisTick={formatPercentTick}
-                      formatYAxisTick={formatPercentTick}
-                      xCountKey="rates-by-race--totals--searches"
-                      yCountKey="rates-by-race--totals--contraband"
-                      xCountLabel={m?.agency_scatter_searches_label?.() ?? "Searches"}
-                      yCountLabel={
-                        m?.agency_scatter_contraband_hits_label?.() ?? "Contraband hits"
-                      }
-                      excludeAgencies={scatterExcludedAgencies}
-                      minStops={500}
-                      sizeByStops={true}
-                      stopsLabel={m?.agency_scatter_total_stops_label?.() ?? "Total stops"}
-                      minCount={25}
-                      minCountKey="rates-by-race--totals--searches"
-                      excludeAboveX={50}
-                      minCountMessage={
-                        m?.agency_scatter_min_searches_note?.() ??
-                          "Requires at least 25 searches to display."
-                      }
-                      note={
-                        m?.agency_scatter_min_searches_note?.() ??
-                          "Requires at least 25 searches to display."
-                      }
-                    />
-                    <AgencyRateScatter
-                      selectedYear={selectedYear}
-                      agencyName={agencyData?.agency ?? data.slug}
-                      title={
-                        m?.agency_scatter_search_vs_hit_nonwhite_heading?.() ??
-                          "Non-white search rate vs contraband hit rate"
-                      }
-                      domainGroup="hits-searches"
-                      showMeanLines={true}
-                      xLabel={
-                        m?.agency_scatter_search_rate_nonwhite_label?.() ??
-                          "Non-white search rate"
-                      }
-                      yLabel={
-                        m?.agency_scatter_hit_rate_nonwhite_label?.() ??
-                          "Non-white hit rate"
-                      }
-                      xMetricKey="rates-by-race--totals--searches-rate"
-                      yMetricKey="rates-by-race--totals--contraband-hit-rate"
-                      xColumn="Non-white"
-                      yColumn="Non-white"
-                      formatXAxisTick={formatPercentTick}
-                      formatYAxisTick={formatPercentTick}
-                      xCountKey="rates-by-race--totals--searches"
-                      yCountKey="rates-by-race--totals--contraband"
-                      xCountColumn="Non-white"
-                      yCountColumn="Non-white"
-                      xCountLabel={
-                        m?.agency_scatter_searches_nonwhite_label?.() ??
-                          "Non-white searches"
-                      }
-                      yCountLabel={
-                        m?.agency_scatter_contraband_hits_nonwhite_label?.() ??
-                          "Non-white contraband hits"
-                      }
-                      excludeAgencies={scatterExcludedAgencies}
-                      minStops={500}
-                      sizeByStops={true}
-                      stopsColumn="Non-white"
-                      stopsLabel="Non-white total stops"
-                      minCount={25}
-                      minCountKey="rates-by-race--totals--searches"
-                      minCountColumn="Non-white"
-                      excludeAboveX={50}
-                      minCountMessage={
-                        m?.agency_scatter_min_searches_note_small?.() ??
-                          "Requires at least 25 searches to display."
-                      }
-                      note={
-                        m?.agency_scatter_min_searches_note_small?.() ??
-                          "Requires at least 25 searches to display."
-                      }
-                    />
-                    <AgencyRateScatter
-                      selectedYear={selectedYear}
-                      agencyName={agencyData?.agency ?? data.slug}
-                      title={
-                        m?.agency_scatter_search_vs_hit_white_heading?.() ??
-                          "White search rate vs contraband hit rate"
-                      }
-                      domainGroup="hits-searches"
-                      showMeanLines={true}
-                      xLabel={
-                        m?.agency_scatter_search_rate_white_label?.() ?? "White search rate"
-                      }
-                      yLabel={m?.agency_scatter_hit_rate_white_label?.() ?? "White hit rate"}
-                      xMetricKey="rates-by-race--totals--searches-rate"
-                      yMetricKey="rates-by-race--totals--contraband-hit-rate"
-                      xColumn="White"
-                      yColumn="White"
-                      formatXAxisTick={formatPercentTick}
-                      formatYAxisTick={formatPercentTick}
-                      xCountKey="rates-by-race--totals--searches"
-                      yCountKey="rates-by-race--totals--contraband"
-                      xCountColumn="White"
-                      yCountColumn="White"
-                      xCountLabel={m?.agency_scatter_searches_white_label?.() ?? "White searches"}
-                      yCountLabel={
-                        m?.agency_scatter_contraband_hits_white_label?.() ??
-                          "White contraband hits"
-                      }
-                      excludeAgencies={scatterExcludedAgencies}
-                      minStops={500}
-                      sizeByStops={true}
-                      stopsColumn="White"
-                      stopsLabel="White total stops"
-                      minCount={25}
-                      minCountKey="rates-by-race--totals--searches"
-                      minCountColumn="White"
-                      excludeAboveX={50}
-                      minCountMessage={
-                        m?.agency_scatter_min_searches_note_small?.() ??
-                          "Requires at least 25 searches to display."
-                      }
-                      note={
-                        m?.agency_scatter_min_searches_note_small?.() ??
-                          "Requires at least 25 searches to display."
-                      }
-                    />
-                  </div>
-                </div>
-              </div>
-            </div>
+            <ScatterSection
+              {years}
+              {selectedYear}
+              agencyName={agencyData?.agency ?? data.slug}
+              excludeAgencies={scatterExcludedAgencies}
+              {formatPercentTick}
+              on:yearselect={(e) => selectYear(e.detail.year, e.detail.source)}
+            />
           </div>
         {/if}
       </article>
